third_party/WebKit/LayoutTests/css3/device-adapt/opera/syntax-002-expected.txt


FAIL CSS Test: @viewport syntax - unknown descriptor assert_equals: expected 500 but got 800
